+++ src/components/components.ts
@@ -1,4 +1,4 @@
-/* Web components. File generated at Wed, 19 Mar 2025 22:32:46 GMT. */
+/* Web components. File generated at Fri, 21 Mar 2025 03:35:17 GMT. */
 export * from './collapsible-content-panel/collapsible-content-panel.component';
 export * from './fixed-scroll-panel/fixed-scroll-panel.component';
 export * from './command-prompt/command-prompt.component';