+++ src/outputFunction.ts
@@ -1,8 +1,8 @@
-import PlotEngine from './PlotEngine';
+import { PlotEngine } from './PlotEngine';
 
 const insertOutput = { type: '' };
 const outputFunction: { [k: string]: Function } = {
     ...PlotEngine.outputFunction,
 };
 export { outputFunction, insertOutput };
-export default outputFunction;
+export default { outputFunction, insertOutput };